Fix integration tests for Ansible 2.19.
This commit makes the necessary changes to run integration tests with
Ansible 2.16 through Ansible 2.19. Because Ansible 2.19 does not work
with any Ubuntu release that supports Percona Server 5.7, this commit
upgrades the `percona-server-client` to version 8.0, which is currently
the oldest supported release of Percona Server.

This commit was large enough that it didn't fix the Ansible 2.19
warnings, just the errors.