Fix integration tests for Ansible 2.19. (#178)
* Fix integration tests for Ansible 2.19.

This commit makes the necessary changes to run integration tests with
Ansible 2.16 through Ansible 2.19. Because Ansible 2.19 does not work
with any Ubuntu release that supports Percona Server 5.7, this commit
upgrades the `percona-server-client` to version 8.0, which is currently
the oldest supported release of Percona Server.

This commit was large enough that it didn't fix the Ansible 2.19
warnings, just the errors.

* Coerce types for python3.12+ansible2.17.

Some tests in `proxysql_global_variables` compared strings to integers,
which doesn't seem to work in all permutations of python+ansible. By
coercing the comparisons to strings, the issue is prevented.

* Create systemd overrides for integration tests.

In Ubuntu 24.04, the combination of `PrivateDevices` and
`RestrictAddressFamilies` was leading to a need for the `CAP_SYS_ADMIN`
capability within the docker container. These stanzas within systemd are
intended to sandbox processes, reducing the attack surface of units run
within it.

While I think both of these systemd stanzas may make sense for
production deployments, I am not sure they make sense on ephemeral test
containers --- especially if it requires additional system privileges by
running the container in privileged mode.

This commit adds a systemd override to remove both systemd
configurations for the unit.