Add `proxysql_mysql_hostgroup_attributes`. (#180)
* Add `proxysql_mysql_hostgroup_attributes`.

The `[mysql_hostgroup_attributes]` table allows for per-hostgroup
overrides, which change the behaviour of a specific hostgroup.

For example, setting the `max_num_online_servers` value for a host group
will change ProxySQL's behaviour when the number of `ONLINE` servers
exceed this value. Similarly, multiplexing may be enabled or disabled on
a per-hostgroup basis.

This PR adds support for managing `mysql_hostgroup_attributes`.

[mysql_hostgroup_attributes]: https://proxysql.com/documentation/main-runtime/#mysql_hostgroup_attributes

* More documentation for `hostgroup_attributes`.

This commit adds additional documentation for the
`proxysql_mysql_hostgroup_attributes` plugin.

Co-authored-by: Andrew Klychkov <aaklychkov@mail.ru>

---------

Co-authored-by: Andrew Klychkov <aaklychkov@mail.ru>